backend/requirements.txt
flask==3.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
mongoengine==0.27.0
werkzeug==3.0.1
google-cloud-storage==2.14.0
Flask-JWT-Extended==4.6.0
certifi==2023.11.17
